Move connector interfaces to the spi package
