+++ core-tck/src/main/java/org/eclipse/microprofile/reactive/messaging/tck/connector/DummyConnector.java
@@ -21,8 +21,8 @@ package org.eclipse.microprofile.reactive.messaging.tck.connector;
 import org.eclipse.microprofile.config.Config;
 import org.eclipse.microprofile.reactive.messaging.Message;
 import org.eclipse.microprofile.reactive.messaging.MessagingProvider;
-import org.eclipse.microprofile.reactive.messaging.connector.IncomingConnectorFactory;
-import org.eclipse.microprofile.reactive.messaging.connector.OutgoingConnectorFactory;
+import org.eclipse.microprofile.reactive.messaging.spi.IncomingConnectorFactory;
+import org.eclipse.microprofile.reactive.messaging.spi.OutgoingConnectorFactory;
 import org.eclipse.microprofile.reactive.streams.operators.PublisherBuilder;
 import org.eclipse.microprofile.reactive.streams.operators.ReactiveStreams;
 import org.eclipse.microprofile.reactive.streams.operators.SubscriberBuilder;
